+++ FlowCrypt/Functionality/Mail Provider/MessageOperations Provider/MessageOperationsProvider.swift
@@ -16,5 +16,6 @@ protocol MessageOperationsProvider {
     func archiveMessage(message: Message, folderPath: String) async throws
     func markAsUnread(message: Message, folder: String) async throws
     func markAsRead(message: Message, folder: String) async throws
+    func emptyFolder(path: String) async throws
     func batchDeleteMessages(identifiers: [String], from folderPath: String?) async throws
 }
